% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{fs_degf}
\alias{fs_degf}
\title{Degrees of freedom of a design}
\usage{
fs_degf(design)
}
\arguments{
\item{design}{An `fsvy` object.}
}
\description{
Number of PSUs minus number of strata among rows in the current domain,
as `survey::degf()`.
}
