% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapidsvy-package.R
\docType{package}
\name{rapidsvy-package}
\alias{rapidsvy}
\alias{rapidsvy-package}
\title{rapidsvy: rapid design-based estimation for complex surveys}
\description{
Taylor linearisation for stratified one-stage cluster designs (PSUs
sampled with replacement, optional finite population correction). All
domains of a `group_by()` are estimated in one multithreaded pass, so the
cost does not grow with the number of groups the way `survey::svyby()` and
`srvyr` do.
}
\details{
Number of threads: `options(rapidsvy.threads = n)`; default is all cores
minus one.
}
\author{
\strong{Maintainer}: Ahmad Watsiq Maula \email{ahmad.watsiq.m@mail.ugm.ac.id}

Authors:
\itemize{
  \item Ahmad Watsiq Maula \email{ahmad.watsiq.m@mail.ugm.ac.id}
}

}
\keyword{internal}
