% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{tidy.fs_glm}
\alias{tidy.fs_glm}
\title{Tidy an fs_glm fit}
\usage{
\method{tidy}{fs_glm}(x, conf.int = FALSE, conf.level = 0.95, exponentiate = FALSE, ...)
}
\arguments{
\item{x}{An `fs_glm` object.}

\item{conf.int}{Add confidence intervals.}

\item{conf.level}{Confidence level.}

\item{exponentiate}{Exponentiate estimates and intervals (odds/rate ratios).}

\item{...}{Unused.}
}
\description{
Tidy an fs_glm fit
}
